Backend/requirements.txt
alembic==1.13.2
beautifulsoup4==4.12.3
bcrypt==4.1.3
celery==5.4.0
dnspython==2.6.1
fastapi==0.115.2
httpx==0.27.2
jinja2==3.1.4
langchain==1.2.5
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.2.0
langchain-groq==0.2.0
langchain-mistralai==0.2.0
langchain-cohere==0.3.0
langchain-community==0.3.0
lxml==5.3.0
neo4j==5.25.0
openai==1.50.0
anthropic==0.34.0
passlib==1.7.4
playwright==1.48.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
redis==5.1.1
sqlalchemy==2.0.36
uvicorn[standard]==0.32.0
asyncpg==0.30.0
weasyprint==62.3
